upgrade to antora 3
