+++ .drone.star
@@ -19,6 +19,9 @@ def main(ctx):
     deployment_branch = default_branch
     pdf_branch = default_branch
 
+    # Environment variables needed to generate the search index are only provided in the docs repo in .drone.star
+    # Also see the documentation for more details.
+
     return [
         checkStarlark(),
         build(ctx, environment, latest_version, deployment_branch, base_branch, pdf_branch),
@@ -86,33 +89,17 @@ def build(ctx, environment, latest_version, deployment_branch, base_branch, pdf_
             {
                 "name": "docs-deps",
                 "pull": "always",
-                "image": "owncloudci/nodejs:16",
+                "image": "owncloudci/nodejs:18",
                 "commands": [
                     "yarn install",
                 ],
             },
             {
-                "name": "docs-validate",
-                "pull": "always",
-                "image": "owncloudci/nodejs:16",
-                "commands": [
-                    "yarn validate --fetch",
-                ],
-            },
-            {
                 "name": "docs-build",
                 "pull": "always",
-                "image": "owncloudci/nodejs:16",
+                "image": "owncloudci/nodejs:18",
                 "commands": [
-                    "yarn antora --fetch --attribute format=html",
-                ],
-            },
-            {
-                "name": "docs-pdf",
-                "pull": "always",
-                "image": "owncloudci/asciidoctor:latest",
-                "commands": [
-                    "bin/makepdf -m",
+                    "yarn antora",
                 ],
             },
             {
@@ -157,29 +144,30 @@ def build(ctx, environment, latest_version, deployment_branch, base_branch, pdf_
                     ],
                 },
             },
-            {
-                "name": "upload-pdf",
-                "pull": "always",
-                "image": "plugins/s3-sync",
-                "settings": {
-                    "bucket": "uploads",
-                    "endpoint": from_secret("docs_s3_server"),
-                    "access_key": from_secret("docs_s3_access_key"),
-                    "secret_key": from_secret("docs_s3_secret_key"),
-                    "path_style": "true",
-                    "source": "pdf_web/",
-                    "target": "/pdf/%s" % environment,
-                },
-                "when": {
-                    "event": [
-                        "push",
-                        "cron",
-                    ],
-                    "branch": [
-                        pdf_branch,
-                    ],
-                },
-            },
+            # we keep uploading pdf for future reenabling
+            #{
+            #    "name": "upload-pdf",
+            #    "pull": "always",
+            #    "image": "plugins/s3-sync",
+            #    "settings": {
+            #        "bucket": "uploads",
+            #        "endpoint": from_secret("docs_s3_server"),
+            #        "access_key": from_secret("docs_s3_access_key"),
+            #        "secret_key": from_secret("docs_s3_secret_key"),
+            #        "path_style": "true",
+            #        "source": "pdf_web/",
+            #        "target": "/pdf/%s" % environment,
+            #    },
+            #    "when": {
+            #        "event": [
+            #            "push",
+            #            "cron",
+            #        ],
+            #        "branch": [
+            #            pdf_branch,
+            #        ],
+            #    },
+            #},
             {
                 "name": "notify",
                 "pull": "if-not-exists",
