+++ es-docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3'
+services:
+  elasticsearch:
+    image: elasticsearch:7.17.15
+    ports:
+      - 9200:9200
+      - 9300:9300
+    environment:
+      - discovery.type=single-node
+      - xpack.security.enabled=false
