+++ site.yml
@@ -31,9 +31,6 @@ asciidoc:
     experimental: true
     # Antora pagination (prev page, next page)
     page-pagination: true
-    # Custom attribute page-component-build-list
-    # Only embed named opengraph images in docs-ui via partials/head-meta-opengraph.hbs
-    page-component-build-list: 'docs, server, ocis, webui, user, desktop, ios-app, android'
 #   common
     docs-base-url: 'https://doc.owncloud.com'
     oc-complete-base-url: 'https://download.owncloud.com/server/stable'
@@ -43,9 +40,13 @@ asciidoc:
     oc-central-url: 'https://central.owncloud.org'
     oc-support-url: 'https://owncloud.com/support'
 #   branded
-    latest-branded-version: 'next'
-    previous-branded-version: 'next'
+    latest-branded-version: 'next@'   # do not change, soft set, correctly defined via antora.yml
+    previous-branded-version: 'next@' # do not change, soft set, correctly defined via antora.yml
   extensions:
-    - ./lib/extensions/tabs.js
-    - ./lib/extensions/remote-include-processor.js
-    - ./node_modules/asciidoctor-kroki/src/asciidoctor-kroki.js
+    - ./ext-asciidoc/tabs.js
+    - ./ext-asciidoc/remote-include-processor.js
+    - asciidoctor-kroki
+
+antora:
+  extensions:
+    - ./ext-antora/generate-index.js
